refactor(audit): add api/web/audit
