+++ src/dashboard/apigateway/apigateway/apis/open/stage/views.py
@@ -24,8 +24,8 @@ from rest_framework import status, viewsets
 from apigateway.apis.open.stage import serializers
 from apigateway.apis.open.stage.serializers import StageSLZ
 from apigateway.apps.audit.constants import OpObjectTypeEnum, OpStatusEnum, OpTypeEnum
-from apigateway.apps.audit.utils import record_audit_log
 from apigateway.biz.released_resource import ReleasedResourceDataHandler
+from apigateway.common.audit.shortcuts import record_audit_log
 from apigateway.common.permissions import GatewayRelatedAppPermission
 from apigateway.core.constants import StageStatusEnum
 from apigateway.core.models import Stage
