+++ src/dashboard/apigateway/apigateway/apis/web/audit/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # TencentBlueKing is pleased to support the open source community by making
 # 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
@@ -16,10 +15,3 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-from django.urls import path
-
-from apigateway.apps.audit.views import AuditEventLogViewSet
-
-urlpatterns = [
-    path("logs/", AuditEventLogViewSet.as_view({"get": "list"}), name="audit.logs.list"),
-]
